modified ush/ioda/bufr2ioda/bufr2ioda_satwind_amv_seviri.py: lat2 references its own dtype in output to netcdf, instead of referencing dtype of lat

+++ ush/ioda/bufr2ioda/bufr2ioda_satwind_amv_seviri.py
@@ -320,7 +320,7 @@ def bufr_to_ioda(config, logger):
                 .write_data(lon2)
 
             # Latitude
-            obsspace.create_var('MetaData/latitude', dtype=lat.dtype, fillval=lat2.fill_value) \
+            obsspace.create_var('MetaData/latitude', dtype=lat2.dtype, fillval=lat2.fill_value) \
                 .write_attr('units', 'degrees_north') \
                 .write_attr('valid_range', np.array([-90, 90], dtype=np.float32)) \
                 .write_attr('long_name', 'Latitude') \
